Demo Project

+++ src/main/java/com/curdexample/crudexample/Services/CategoryService.java
@@ -3,11 +3,11 @@ package com.curdexample.crudexample.Services;
 import com.curdexample.crudexample.Exception.ResourceNotFoundException;
 import com.curdexample.crudexample.dao.CategoryDao;
 import com.curdexample.crudexample.entities.Category;
+import com.curdexample.crudexample.entities.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 
 @Service
@@ -15,30 +15,63 @@ public class CategoryService implements CategorySeviceInter {
     @Autowired
     private CategoryDao categoryDao;
 
+    /**
+     * This is the service method for getting  all data of category
+     */
     @Override
     public List<Category> getCategory() {
         return categoryDao.findAll();
     }
 
+    /**
+     * This is the service method for getting category  data of  given id
+     *
+     * @param categoryId
+     */
     @Override
     public Category getCategory(int categoryId) {
-        return categoryDao.findById(categoryId).get();
+        if (categoryDao.findById(categoryId).isPresent()) {
+            return categoryDao.findById(categoryId).get();
+        }
+        throw new ResourceNotFoundException();
     }
 
+    /**
+     * This is the service method for saving category data
+     *
+     * @param category
+     */
     @Override
     public Category addCategory(Category category) {
-        categoryDao.save(category);
-
-        return category;
+        try {
+            categoryDao.save(category);
+            return category;
+        } catch (Exception e) {
+            throw new ResourceNotFoundException();
+        }
     }
 
+    /**
+     * This is the service method for deleting category data
+     *
+     * @param categoryId
+     */
     @Override
     public String deleteCategory(int categoryId) {
-        categoryDao.deleteCategoryById(categoryId);
-        return "deleted";
-
+        if (categoryDao.findById(categoryId).isPresent()) {
+            categoryDao.deleteCategoryById(categoryId);
+            return "deleted";
+        } else {
+            throw new ResourceNotFoundException();
+        }
     }
 
+    /**
+     * This is the service method for updating category data
+     *
+     * @param category
+     * @param categoryId
+     */
     @Override
     public Category updateCategory(Category category, int categoryId) {
         Category update;
@@ -56,5 +89,21 @@ public class CategoryService implements CategorySeviceInter {
             throw new ResourceNotFoundException();
         }
         return update;
+    }
+
+    @Override
+    public Category CheckNameCategory(Category category) {
+        if (category.getCategoryName() == null) {
+            throw new IllegalArgumentException("Category name is null");
+        } else {
+            int len = category.getCategoryName().length();
+            String s = category.getCategoryName();
+            for (int i = 0; i < len; i++) {
+                if (Character.isLetterOrDigit(s.charAt(i)) == false) {
+                    throw new RuntimeException("ProductName should contain Letters only");
+                }
+            }
+            return category;
+        }
     }
 }
